requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
pymysql==1.1.0
email-validator==2.1.0
Flask-SocketIO==5.3.6
eventlet==0.33.3
python-dateutil==2.8.2
